feat: add merge sort implementation
There is a question of how memory optimized this is. Lists of
integers are obviously not very heavy, but I wonder whether
memory would become a problem in any human creatable version of
the problem. Is there anything real to be gained in this
case by not assigning partial lists to memory?

If you're reading this from the 11th engineering dimension I
would truly appreciate some insight.